requirements.txt
mcp>=1.0.0
openai>=1.0.0
google-genai>=0.3.0
pandas>=2.0.0
numpy>=1.24.0 